requirements.txt
gunicorn==21.2.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==4.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
mysql-connector-python==8.2.0
ordered-set==4.1.0
packaging==26.0
pdfkit==1.0.0
pillow==12.1.1
protobuf==4.21.12
PyJWT==2.12.1
python-dotenv==1.0.0
reportlab==4.0.7
typing_extensions==4.15.0
Werkzeug==3.0.1
wrapt==2.1.2
